last didnt work, reverting back to long to test

+++ src/main/java/com/wahine/vippsstats/model/Customer.java
@@ -8,7 +8,7 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "customer", schema = "public")
 public class Customer {
-    private int customerId;
+    private long customerId;
     private String firstName;
     private String lastName;
     private String emailId;
@@ -33,7 +33,7 @@ public class Customer {
         this.emailId = emailId;
     }
 
-    public Customer(int customerId, String firstName, String lastName, String emailId) {
+    public Customer(long customerId, String firstName, String lastName, String emailId) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.emailId = emailId;
@@ -42,10 +42,10 @@ public class Customer {
     @Id
     //@GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "customer_id", nullable = false)
-    public int getCustomerId() {
+    public long getCustomerId() {
         return customerId;
     }
-    public void setCustomerId(int customerId) {
+    public void setCustomerId(long customerId) {
         this.customerId = customerId;
     }
